修正函数setByteArray(String,int[])的名字为setIntArray

+++ src/main/java/cc/bukkitPlugin/commons/nmsutil/nbt/wapper/NBTWCompound.java
@@ -190,7 +190,7 @@ public class NBTWCompound extends NBTWBase<Map>{
         return tNBTWTag instanceof NBTWIntArray?((NBTWIntArray)tNBTWTag).getValue():new int[0];
     }
 
-    public void setByteArray(String pKey,int[] pValue){
+    public void setIntArray(String pKey,int[] pValue){
         this.set(pKey,new NBTWIntArray(pValue));
     }
